# .schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "runtime",
    "description": "The spiced runtime",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1.2.0-unstable"
  },
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Local development server. Configure with `--http`."
    }
  ],
  "paths": {
...
    "/v1/mcp/event": {
      "post": {
        "tags": [
          "mcp"
        ],
        "summary": "Send message to MCP server",
        "description": "Send message to the MCP endoint, for a given session.",
        "operationId": "mcp_event",
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Message accepted. Response will stream via SSE."
          },
          "404": {
            "description": "Session not found. No active session for the given `session_id`."
          },
          "413": {
            "description": "Payload too large. Maximum allowed size is 4MB."
          },
          "500": {
            "description": "Internal server error. An unexpected issue occurred."
          }
        }
      }
    },
...
    "/v1/models/{name}/predict": {
      "get": {
        "tags": [
          "AI"
        ],
        "summary": "ML Prediction",
        "description": "Make a ML prediction using a specific model.",
        "operationId": "get_model_predict",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "The name of the model to make a prediction with.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prediction made successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                },
                "example": {
                  "status": "Success",
                  "model_name": "my_model_name",
                  "model_version": "1.0",
                  "prediction": [
                    0.45,
                    0.5,
                    0.55
                  ],
                  "duration_ms": 123
                }
              }
            }
          },
          "400": {
            "description": "Invalid request to the model",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                },
                "example": {
                  "status": "BadRequest",
                  "error_message": "You gave me a bad request :(",
                  "model_name": "my_model_name",
                  "duration_ms": 12
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occurred during prediction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                },
                "example": {
                  "status": "InternalError",
                  "error_message": "Oops, the server couldn't predict",
                  "model_name": "my_model_name",
                  "duration_ms": 12
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/predict": {
      "post": {
        "tags": [
          "AI"
        ],
        "summary": "Batch ML Predictions",
        "description": "Perform a batch of ML predictions, using multiple models, in one request. This is useful for ensembling or A/B testing different models.",
        "operationId": "post_batch_predict",
        "requestBody": {
          "description": "Batch prediction request containing a list of prediction requests for specific models",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchPredictRequest"
              },
              "example": {
                "predictions": [
                  {
                    "model_name": "drive_stats_a"
                  },
                  {
                    "model_name": "drive_stats_b"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Batch predictions completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchPredictResponse"
                },
                "example": {
                  "duration_ms": 81,
                  "predictions": [
                    {
                      "status": "Success",
                      "model_name": "drive_stats_a",
                      "model_version": "1.0",
                      "prediction": [
                        0.45,
                        0.5,
                        0.55
                      ],
                      "duration_ms": 42
                    },
                    {
                      "status": "Success",
                      "model_name": "drive_stats_b",
                      "model_version": "1.0",
                      "prediction": [
                        0.43,
                        0.51,
                        0.53
                      ],
                      "duration_ms": 42
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error occurred during batch prediction",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                },
                "example": "An unexpected error occurred while processing batch predictions"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchPredictRequest": {
        "type": "object",
        "properties": {
          "predictions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PredictRequest"
            },
            "description": "The list of prediction requests, each specifying the model to use for the prediction"
          }
        }
      },
      "BatchPredictResponse": {
        "type": "object",
        "required": [
          "predictions",
          "duration_ms"
        ],
        "properties": {
          "duration_ms": {
            "type": "integer",
            "minimum": 0
          },
          "predictions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PredictResponse"
            }
          }
        }
      },
...
      "PredictRequest": {
        "type": "object",
        "required": [
          "model_name"
        ],
        "properties": {
          "model_name": {
            "type": "string"
          }
        }
      },
      "PredictResponse": {
        "type": "object",
        "required": [
          "status",
          "model_name",
          "duration_ms"
        ],
        "properties": {
          "duration_ms": {
            "type": "integer",
            "description": "The time taken to complete the prediction (in milliseconds)",
            "minimum": 0
          },
          "error_message": {
            "type": [
              "string",
              "null"
            ],
            "description": "The error message if the request failed"
          },
          "model_name": {
            "type": "string",
            "description": "The name of the model used for the prediction"
          },
          "model_version": {
            "type": [
              "string",
              "null"
            ],
            "description": "The version of the model used"
          },
          "prediction": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "number",
              "format": "float"
            },
            "description": "The prediction result, typically an array of floats"
          },
          "status": {
            "$ref": "#/components/schemas/PredictStatus",
            "description": "The status of the prediction"
          }
        }
      },
      "PredictStatus": {
        "type": "string",
        "enum": [
          "Success",
          "BadRequest",
          "InternalError"
        ]
      },
...
    }
  },
  "security": [
    {
      "api_key": []
    }
  ]
}